Fixed compiler warnings (#35)
* Update sevenzip.c

* Update hpdf_streams.c

* Update hpdf_streams.c

* Update hpdf_image_png.c

* Update base.c

* Update dbcopy.c

* Update dbdelim.c

* Update dbsort.c

* Update filesys.c

* Update gzread.c

* Update gzwrite.c

* Update hbpp.c

* Update hbppcomp.c

* Update hbppcore.c

* Update pragma.c

* Update mxml_fil.c

* Update xemit.c

* Update xpatchi.c

* Update xutils.c

* Update ChangeLog

+++ contrib/hbxdiff/xemit.c
@@ -35,7 +35,7 @@ static long xdl_get_rec(xdfile_t *xdf, long ri, char const **rec) {
 
 
 static int xdl_emit_record(xdfile_t *xdf, long ri, char const *pre, xdemitcb_t *ecb) {
-	long size, psize = strlen(pre);
+	long size, psize = ( long ) strlen(pre);
 	char const *rec;
 
 	size = xdl_get_rec(xdf, ri, &rec);
@@ -132,4 +132,3 @@ int xdl_emit_diff(xdfenv_t *xe, xdchange_t *xscr, xdemitcb_t *ecb,
 
 	return 0;
 }
-
