+++ source/rtl/zlib/gzread.c
@@ -310,7 +310,7 @@ local z_size_t gz_read( gz_statep state, voidp buf, z_size_t len )
         /* set n to the maximum amount of len that fits in an unsigned int */
         n = ( unsigned int ) -1;
         if (n > len)
-            n = len;
+            n = ( unsigned int ) len;
 
         /* first just try copying data from the output buffer */
         if (state->x.have) {
@@ -388,7 +388,7 @@ int ZEXPORT gzread( gzFile file, voidp buf, unsigned len )
     }
 
     /* read len or fewer bytes to buf */
-    len = gz_read(state, buf, len);
+    len = ( unsigned int ) gz_read(state, buf, len);
 
     /* check for an error */
     if (len == 0 && state->err != Z_OK && state->err != Z_BUF_ERROR)
@@ -455,7 +455,7 @@ int ZEXPORT gzgetc( gzFile file )
     }
 
     /* nothing there -- try gz_read() */
-    ret = gz_read(state, buf, 1);
+    ret = ( int ) gz_read(state, buf, 1);
     return ret < 1 ? -1 : buf[0];
 }
